1-data-pipeline/requirements.txt
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.10.0
yfinance>=0.2.48
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
